pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "django-rest-framework-keycloak-auth"
version = "0.1.0"
dependencies = [
    "django>=5.1",
    "djangorestframework>=3.15",
    "python-keycloak>=4.6"
]
description = "A Django app to use SSO with Keycloak on your REST API."
readme = "README.md"
requires-python = ">= 3.10"
authors = [
    {name = "Julien Deslandes", email = "noklm@proton.me"},
]
license = {file = "LICENSE"}
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 5.1",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]

[project.urls]
Homepage = "https://github.com/Noklm/django-rest-framework-keycloak-auth"